requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
redis>=5.0.0
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.0
